Move testing files to proper place
