+++ tests/torchtune/_cli/test_ls.py
@@ -10,7 +10,7 @@ import sys
 
 from recipes import list_configs, list_recipes
 
-from tests._cli.common import TUNE_PATH
+from tests.torchtune._cli.common import TUNE_PATH
 
 from torchtune._cli.ls import _NULL_VALUE
